Address PR comments of renaming
