+++ examples/load_sghmc.py
@@ -2,9 +2,7 @@ from sghmc.modules.classifier import Classifier
 from uqlib import load_optimizer_param_to_model
 import torch
 
-ckpt_path = (
-    "/home/paperspace/Projects/bayes-lms/examples/lightning_logs/checkpoints/last.ckpt"
-)
+ckpt_path = None
 
 model = Classifier()
 groups = torch.load(ckpt_path)["save_params"][9]
